refactor(scm): address PR review notes on test seams
Move checkout_pull_request to tests/scm/support.py (protocol-test seam only),
inline ctx.scm instead of resolve_scm_provider, and add GitHub adapter
validate_provider coverage.
